+++ modules/iam-assumable-role/main.tf
@@ -117,6 +117,15 @@ data "aws_iam_policy_document" "assume_role_with_mfa" {
         values   = local.role_sts_externalid
       }
     }
+
+    dynamic "condition" {
+      for_each = var.role_requires_session_name ? [1] : []
+      content {
+        test     = "StringEquals"
+        variable = "sts:RoleSessionName"
+        values   = var.role_session_name
+      }
+    }
   }
 }